requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
python-telegram-bot==20.7
httpx==0.25.2
celery==5.3.4
redis==5.0.1
